admins: create new group sitemaps-admins for sitemaps uploaders
Add a new admin group for people who upload files to
sitemaps.wikimedia.org.

Add the current members of the perf-team to add it. Not using
that group because it includs full root and was originally for
another purpose.

The members can overlap with perf-team but don't have to in the future.

Bug: T202910
Change-Id: I97b190e2f37d022b68434510fa512908b2624196