labstore: Allow override for load monitoring
Since labstore1004/5 now have a kernel that likes to eat more load,
and we do not truly want to be paged constantly for what is normal,
this should keep the monitors consistent for other labstores while
setting the ones on these much higher.  They should be able to handle it.

Bug: T202323
Change-Id: I9385e1e68a4ef8fdfa432efe023c598ad8c6e3d8