profile::analytics::cluster::packages::statistics: add lynx
Sometimes it happens that yarn.wikimedia.org is not able to
proxy to each analytics10XX Yarn HTTP server to inspect logs
(we are still trying to come up with a httpd config able to do
that but it is not trivial). As interim step, having lynx on
the stat machines can surely help developers.

Change-Id: I52c70825b4ba0acb919e68239ad475f1c87d6ad6