rsync::server: add parameter to use IPv6
Following up to the comments on I5603e60132fc06bab,
make it more obvious how to enable IPv6 and what the
behaviour actually is by adding a parameter called
use_ipv6 which sets the address to ''.

As confirmed by our tests, when address is left blank
rsync listens on :::873 in addition to 0.0.0.0:873.

Change-Id: I23dce2d94f3e83fa8d90fb7b14e37e10da50d94f