block_sync: Small improvement to the drbd backup script
The backup script works well.  However, on failure, it doesn't "fail".
This appears to be because regardless of the status of the bdsync command
the python script will exit successfully.  To fix that, this performs
a sys.exit with the status code of the bdsync command, which will still
run the finally block, cleaning up the locks.  This will cause systemd
to enter a failed state and alert on icinga.

Bug: T171394
Change-Id: Ic3e8079f3ffd7ae9d216fe2425152fe0a172bf17