conftool: add class for writing to state to file
This is used on the puppetmasters in order to retrieve via confd
variables that we want to later use in our manifests. This is more
scalable that the older approach (using the conftool parser function) as
we will have only one, proven method to sync data from etcd, but is
still not ideal. Should only be used for things that are fully
configured via puppet like cronjobs and alerts.

Also:
* add a 'mediawiki::state' puppet function that allows retrieving state
  data for MediaWiki.
* add a deprecation notice to the parser function conftool()
* Add a mock file for the puppet compiler to consume

Change-Id: I71a5508c25250f851ace3a3f842b485b9d561837