profile::mediawiki::maintenance: depend on mediawiki config, not hiera
In normal operating conditions, we want to run MediaWiki maintenance
jobs on the maintenance server which is in the MW primary datacenter. So
start using the same data again.

Please note that this only allows one maintenance server per dc and this
limitation needs to be solved sooner than later - maybe by moving the
MediaWiki jobs to kubernetes.

Change-Id: I6c67e6214a6a1b010cdf4858b891ec9877cc889e