nfs-exportd: remove subtree_check from project exports
subtree_check is not actually needed to export subtrees, since we have fsids.
It is needed if we want to absolutely ensure against a particular type
of attack to see files in a different subtree of the filesystem.  It is
a sort of brute force/guessing type attack, so the risk isn't high.
On the other hand, the cost of subtree_check is quite high in resources,
and I consider it a likely cause of our load problems on the labstores.

Bug: T203254
Change-Id: I3a73041d5f6de787e9c5d98fcbef3d86493bfdec