admins: clean up duplicate membership for mbsantos
mbsantos was first added to the "ldap_only" group
with a commented key to be later added to the section
for actual shell access.

But he wasn't removed from the ldap_only group at that point
so now this is duplicate and the commented key can be removed.

See further up in the file he has active shell access.

Bug: T197237
Change-Id: I87c01240f196b87ae75e5cc8f2765c67a7b72926