icinga::web: install libapache2-mod-php if on stretch
The current Icinga class tries to run 'a2enmod php7.0'
to enable the Apache PHP module but fails because
libapache2-mod-php is not installed.

Previously the package got pulled in by the apache puppet
class and then Icinga was switched to use the httpd module
instead where this doesn't happen automatically.

Require the right package is installed and unify code
with the existing "if stretch" for the php-gd package which is
already there.

Bug: T202782
Change-Id: I5de44c4836d7dc5a8a9c97229294787c36f609dc