labstore: Change tcp buffer settings
There are a few things going on with the NFS servers.  The scheduler
doesn't show any obvious problems (though we need perf installed),
but it is clear that nfsd threads end
up temporarily in uninterruptible sleep quite often (which is the high
load numbers we are seeing) and the tcp buffers aren't meeting expectations
of clients (probably from very large reads and writes).  The buffers could
be causing retransmits and such to happen more, thus causing more concurrent
state "D"s on the NFSd threads.  If needed, changing qdisc settings and such
might come next.

Going to install the kernel perf tools soon as well.

Bug: T203254
Change-Id: I0517c61bf47dcf72acbaf8ac76b2d015faa21626