icinga: make the apache virtual host name configurable
Make it possible to configure a different virtual
host name in Apache rather than hardcoding
icinga.wikimedia.org.

This will allow us to set something like "icinga-stretch" for the
new server and use the web UI as normal without having to manually
hack the entire Letsencrypt/TLS setup. We have HSTS enabled
so you can't even add an exception for a mismatching cert in a browser.

The old method of just hacking local /etc/hosts for testing isn't
going to work here due to that. And i don't want to disable puppet
and manually make it http. This seems nicer as it also brings the
module closer to being reused.

In Hiera, set it to icinga.wikimedia.org on einsteinium and tegmen
as hardcoded before and set it to icinga-stretch.wikimedia.org
on icinga1001. Later switch it over.

Bug: T202782
Change-Id: I452aa9623fd9fd2eebaee53a955bdcb7863568db