tor: ensure libzstd1 is installed and required if on stretch
Despite the existing code to explicitly add the thirdparty repo
and require an apt-get update, we still got only the distro package
version installed.

Manually running an apt-get upgrade results in:

The following packages have been kept back:
  tor

It's held back because it would require libzstd1:

'apt-get -s install tor' or 'apt-get -s --with-new-pkgs upgrade'
show that it wants to install this NEW package and then
upgrade tor.

The following additional packages will be installed:
  libzstd1

So add the package to puppet and install it first, make the tor package
require it so that we don't end up with the wrong package version
when this role gets applied on a new host. Avoid having to do manual steps.

Bug: T196701
Change-Id: Ia3dee59984d574c973942d427c6da5be94bcda23