admins: create shell user for Mathew Onipe
This adds a new user for Mathew. Now he has
the wikimedia.org email and it's used in LDAP.

UID 19981

He provided 2 new keys at T202708#4567812.

Adding a new user to groups is always separate
from creating the user so regardless what the
actual access level will be this is a pre-requisite.

It gives access to bastion hosts (and people.wikimedia.org)
but is limited to that until we add (whichever) group memberships.

Bug: T202708
Change-Id: If04918c01af54f50c982ba8d72a52fd4509c9434