monitoring: enable using notes_url with grafana_alert
Passthrough the $notes_url parameter to monitoring::service
to allow using it with grafana_alerts.

This allows us to add additional URLs to Icinga services
such as links to tickets or runbooks.

One use-case is adding a link to T203485 to the performance
team grafana alerts to tell us to ignore them.

Bug: T197873
Bug: T203485
Change-Id: I9da86d026550954f9240d0379879ae8e016cfeb9