cloudvps: nova-network: also distribute default route in dnsmasq
RFC3442 defines that a DHCP client receiving both a classless-static-route
and a standard route option should just use the former and ignore the later.

In order to push 2 static routes to the clients (default gw + compat net),
let's simply put the 2 in the clasless-static-route option.

While at it, rename the hiera key to something more meaningfull.

Bug: T202636
Change-Id: Id6de7d0ab77f8c5a2e117f552f1cffab8d6770e7
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>