icinga: disable notifications for hosts using role(test)
We are already doing this for role(spare::system) but we don't
disable them for role(test).

This means that even test hosts have active alerts for all the checks
added in base, like disk space, DPKG, puppet run etc.

This patch was triggered by:

< bblack> cp1099 is in role test, I don't know why it would alert at all :P

The request for review here is not technical but whether it's agreeable
that hosts defined as "test" would _not_ send notifications (IRC, email.

The hypothesis is that something that is a test can't be
important enough to alert people on the channel.

It would still be in the web UI for people looking closer there.

Change-Id: I769df910d2609082008520ed3182cb975a585eb6