tor_relay: add Bacula backups of Tor keys
In case of disk failure we would be able to restore
a relay with the same fingerprints and keep our flags
if we have backups of the keys and state data.

Unless there is a reason not to put these in Bacula, make
tor_relay hosts backup hosts.

Create a new fileset for tor on Bacula director that includes
both /var/lib/tor/ and /var/lib/tor-instances/.

During the recent migration the latter wasn't synced at first
because /var/lib/tor does not contain /var/lib/tor-instances/
(vs. /var/lib/tor/instances/).

That also partially triggered wanting to have backups now and
the part that the old machine is now decom'ed and we will only
have a single copy of them around again.

Change-Id: Iba08697a8f8b364363c8803811d829abe9244c3e