mediawiki_maintenance: add inactive warning in motd
This is the same mechanism and code used for the existing
MOTD warning on the deployment server.

As requested in T204604 this is also adding that for maintenance
servers.

Template copied and adjusted to maintenance instead of deployment
server.

It's just missing the feature to name the correct hostname to use since
we don't store that in Hiera yet. Not sure if we should add it there
just for this.

So this resorts to saying to connect to the active
data center instead and that's it.

Bug: T204604
Change-Id: I5a3f651eaa9ed9ea2c06f8ea0e36b227b150e817