mariadb: add mwmaint1002 to grants for production-m5
mwmaint1002 (10.64.16.77) is new hardware that will
replace mwmaint1001 (10.64.16.62).

Please add it to the grants in parallel to allow connections
to production-m5.

This is the mediawiki maintenance server in eqiad (cron jobs for mw).
We would like to do this switch before eqiad is the active data center
again because that makes it easier. We don't have to move active crons.

mwmaint1001 was just a temp. replacement for terbium and is a repurposed
app server. mwmaint1002 is actually the new hardware supposed to replace
terbium.

Bug: T201343
Change-Id: I868d86f137b1a2dde87d6d916e67460a4f1ac15c