shinken - Change Puppet thresholds
Modify the "Sum of Puppet errors" service to check for errors in the
past 90 minutes and, considering that Puppet runs every 30min, accept a
threshold of 33% (1 run) as just a warning and 66% (2 runs) as critical.

Also modify it to report 1 datapoint above the treshold as exceeding
the threshold.

This should help to avoid false positives due to transient NFS errors.

Bug: T161898
Change-Id: I0da02253b55409d7023c01a1cad6d6368c554cc0