scap: use mediawiki canaries from codfw
We have switched over to codfw. scap uses dsh files for the list of
canaries and that is still referring to the eqiad hosts.

Update mediawiki-api-canaries and mediawiki-appserver-canaries with
hosts from codfw.

Surely later we would want a list per datacenter and vary the ones being
used by updating scap.cfg. But that patch should be good enough.

Bug: T204907
Change-Id: I5e485fe2a0da60ce442a1221724e61c8e5fb47da