adding nathante to groups
This patchset add the shell user nathante / Nathan TeBlunthuis to the
groups researchers, statistics-privatedata-users and
analytics-privatedata-users groups.
This is second of two patchsets, the first having added the user to the
module.

Please note this should only be merged after checking the related task
for objections, and after the 3 business day wait has ended on
2018-09-24.

Bug: T204790
Change-Id: I2bfed55224b0df96fa777c6919d851b84cc42ad6