make more obvious which rsyncs are automatic
The parameter $auto_sync to rsync::quickdatacopy
determines whether there is also a cron job created
to automatically run the rsync command or not.

If not then it also sets up rsyncd and the ferm rules
and creates the sync command but there is no cron, so a
human has to run it.

Setting it to true is the default.

But since it now repeatedly cost me time to check again whether
a specific one is actually automatic or not and we were not
consistent about using it, i am adding it explictly to
make things obvious.

Bug: T205037
Change-Id: Ieffe2682a6dcf947ff3252122ef2df81af0c074d