cloud: update MediaWiki-Vagrant container start actions
Add a restart of the LXC hosted apache2 process after bringing the
container up. We have noticed that the LXC container sometimes (always?)
starts the apache2 process before the host provided shares have been
mounted. This can lead to apache2 grabbing the inode of the underlying
mount point rather than the mounted directory. This in turn makes
MediaWiki think that there are no configured wikis to display when it
attempts to route to a wiki based on vhost name through the multiwiki
shim.

Change-Id: Ie615fa8ca3f1fe10e2fa7dfc0a021c309bfbe5f0