profile::mariadb::misc::el::sanitization: swap cron with systemd timer
Upgrading the cron sanitization script to run via systemd timer
(and hence get automatic alarming in case of failures, etc..). A big
assumption is that flock is not needed since once the systemd unit
is "started", another "start" will not cause another instance to be
created. If this is not true, flock will be needed.

Bug: T172532
Change-Id: Iac245edd60432e3d6e8dc50e132ca3193f59488e