icinga: make icinga user configurable
It has been a FIXME forever that the puppet manifests handle the user
and group and then also require both, icinga and nagios users, to exist.

Since I224a32cac5364b1 that has been removed if on stretch.

We want to go back to letting the distro package handle the user.

To be able to set a different one in a clean way, make it configurable
and don't hardcode it in all these places.

Bug: T202782
Change-Id: I86ea60f826d3138b02a3a710057811d40b6fc5c0