icinga: make group configurable, make user configurable pt2
follow-up to I86ea60f826d3138

In addition to the user also make the icinga group configurable.

Also add the user variable in some more spots where it was forgotten
in the previous patch.

Note how there is one single spot where a group 'nagios' is hardcoded.
Not touching this here, only replacing where it currently is 'icinga',
so this change should be yet another noop.

Bug: T202782
Change-Id: Icb95d0bc5237188a8c2bd1e03472d8fd542b07b1