Fix: Regenerate map tiles up to zoom level 9 with notify-tilerator-regen
In modules/tilerator/manifests/regen.pp, a cron job is defined to send a
command to tilerator to regenerate map tiles for lower zoom levels.  It
intends to regenerate zoom levels 0-9, but is sending '9' as the
"beforeZoom" parameter.  "beforeZoom" is an exclusive upper bound, so
currently only zoom levels 0-8 are being regenerated. This updates the
command so that zoom level 9 is also regenerated.

Bug: T202201
Change-Id: Ib99cceca2e6e82052ef10d8682ccd3d33b0171de