toolforge: Update toolserver redirect pages
Update the logo and name used on the www.toolserver.org landing page to
refer to Toolforge. Also update links to point to more modern landing
pages for Toolforge and Cloud Services information. Finally add a proper
ErrorDocument directive so that the custom 404 page is actually used
rather than defaulting to the built-in Apache 404 page.

Change-Id: If377cf8a687b784664378ce3397530bb3d2f223b