icinga: pass user/group from profile, change to nagios on icinga1001
Now that the icinga user and group are configurable in the module,
pass the parameters from the profile class, add Hiera lookups,
default to "icinga" and change them to "nagios" on icinga1001 based on
hostname (for now).

This should fix the situation on icinga1001 where "nagios" is used
by the package and puppet does not attempt to manage user/group anymore.

At the same time it ensures einsteinium/tegmen stay unchanged.

Bug: T202782
Change-Id: I5328ef375c8a246337a779db22c5c02b146330ab