icinga: fix more places with hardcoded user name
In event_handlers::raid and monitor::ores we have 2 more
places that hardcode the 'icinga' name that we are changing
back to the default 'nagios'.

Fixes more puppet errors on icinga1001 due to non-existing user
icinga.

Also follow-up to I5328ef375c8a246. The parameter is now coming
from the profile class/Hiera as it should, so there should be no
default here setting it to 'icinga' anymore. The default is with
the profile now.

Bug: T202782
Change-Id: Ie99fa8602b366203a8d7af46c2b90d7c2703902d