Add the analytics-alerts contact to the analytics contact group
The analytics-alerts contact has been created in the private repo.
Since analytics is moving away from alerting about failures via email
to icinga (especially swapping crons with systemd timers) it is necessary
for us to get notified in a reliable way if icinga-vm is not connected
to our IRC channel.

Bug: T172532
Change-Id: Iac990802f7c822089b2850a8c2219ba7f7e2c8e3