labstore: Change the dumps monitoring thresholds
While the CPU-based monitor is more correct for all generic cases,
spikes tend to last a little while on the dumps NFS servers.  Real problems
tend to be much higher, prolonged load spikes on there.  For the sake of
keeping things quieter except where they need to not be quiet, I'm upping
the thresholds.

Change-Id: I811e813d37ff523d98840d2e7e99883d942eee3b