alerting_host/tcpircbot: remove useless role, include only profile
Do not include a role in another role, especially if that role
does nothing besides including a profile.

Just delete that role class and move the include to the alerting_host
role directly.

The only reason not to do this could be if something in cloud or elswhere
uses the tcpircbot role directly but not the alerting_host role.

Change-Id: I50deda26758bac5e922b5641b591ff72c65e2666