Only allow HTTP port for Hue
If $proxy_enabled is set, access is done via the caches for which we
have a stricter Ferm rule. $proxy_enabled seems to have been used for
the transition from thorium to the individually split analytics-tool*
Ganeti instances and given that this transition is done, let's use
the proxied version unconditionally which allows us to remove the
Ferm rule for direct access to the 8888 port (ideally we'd also fix
Hue to only listen on localhost, but that's seperate).

Change-Id: I300922fe4368464087becef441ff3fbc2922f207