parsoid: connect to MediaWiki via https everywhere
Originally, we separated the use of parsoid via http/https to allow to
use different api pools for parsoid requests originated by
changeprop (coming from the inactive DC) and the ones originated by real
users, as changeprop routinely hammered parsoid and the api. Since then,
the issue with changeprop have been fixed, and we want to move to TLS
between services anyways, so convert parsoid to call mediawiki via TLS
everywhere.

This will also spare us a step in the datacenter switchover procedures,
which is a plus.

Change-Id: I60fbc41fa46e0e90dfa121f3b2b1c3277d6a0a26