xhgui::app: add standard and firewall includes to role
Ia3bdaac18ccc removed the test role from tungsten. The test
role included "standard" and "profile::base::firewall" but
the xhgui app role did not. That lead to puppet errors on tungsten:

Invalid relationship: Exec[git_clone_operations/software/xhprof]
{ require => Package[git] }, because Package[git]
doesn't seem to be in the catalog.

Add standard and firewall in the role to fix it.

Change-Id: I671018cdaecae4fabef3c14c1c2d8ad9d59be3af