mw_maintenance: temp hack to avoid duplicate crons on switch to eqiad
mwmaint1001 is being replaced by mwmaint1002. Add a temp hack
to avoid that once we switch back to eqiad crons will be activated
on both hosts.

Make sure that it runs on mwmaint1002 if in eqiad and the existing
ones on mwmaint2001 in codfw keep running while also making the
switch possible without yet another change.

Revert it after the switch to eqiad and mwmaint1002 is succesful.

Bug: T201343
Change-Id: I621b0c81e463659398b5974065140ca7ecf3a7f7