cloud: Enable systemd unit for MediaWiki-Vagrant
Declare a Puppet service resource for the MediaWiki-Vagrant systemd unit
so that Puppet will enable the unit and ensure it is running. Because
this unit is really a one-time task rather than a long running process,
also add a 'RemainAfterExit=yes' clause to the unit so that systemd will
treat it as running after the script exits.

Change-Id: I387ff9546113c9b85b2e1aab7a22bc0f8dc09667