openstack: add case for stretch and newton in client repos
This is really just a placeholder for when we add pinning for stretch,
however, I cannot find any public repo with ocata packages in stretch, but
we are at least using that value for some servers using virtualenvs.

Note: The point of this is to test running stretch as nfs server, and prep
for future stetch upgrades.

Bug: T203254
Change-Id: Iefac2084b9f0300011e976899e2f9106f801db25