Nova: reduce number of worker nodes
m5 is getting overwhelmed with connections, and we don't
have any real evidence that the current number of workers is
necessary.  So, reduce them slightly and we'll see if
there are any ill effects.

In eqiad this should reduce by a third, from 12 to 8.  In
eqiad1 this will reduce by one fourth, from 10 to 8.

Bug: T202889
Bug: T188589
Change-Id: I3949b0f2e6b6be739c57d1d8a4b25da8a37ff1ce