base: do not allow mailman server to NRPE to other hosts for no reason
The IP listed here used to be the IP of neon, the former Icinga
server (f.e. see T82155#897388).

That is the only reason it is listed here. Nowadays that IP is
fermium, the mailman server since it has been reused.

This is a bug, the mailman server doesn't need to be able to NRPE
to all other hosts.

Bug: T202782
Change-Id: Ibc749f790a3b5873f43268a7e1f3f93b2d1ee48d