k8s: Alert on the sum of operation types
For now, we don't have a reason to alert on specific operation types on
a kubelet increasing, but rather on the aggregate. Remove the
operation_type parameter from the sum() in the prometheus query in order
to implement that. This is expected to lower the number of false
positive alerts

Change-Id: I39534cef00c8c112b33d88b5256f0a884e612a00