labs-ip-alias-dump.py: Fix enumerating IPs in Neutron
In nova-network, all IPS are in the network 'public' so that
was just hard-coded.  In Neutron we still only use one
network, but in theory there could be multiple networks
with arbitrary names, so just enumerate each network
regardless of name.

This will only map floating IPs to private IPs within a given
network.  If we added floating IPs in a different new network
this would break; I'm not sure if that's right or wrong
but I'm also not sure that's a case that will ever actually
happen.

Change-Id: I135176f27806c6879eee3a098b2f3076a4ebeef2