cloudvps: metrics: we don't require observerenv
We don't really depend on observerenv but in novaenv.
And that profile is already deployed in control nodes by means of keystone
profiles. I think there is no need to be explicit about a dependency because
I don't see a future in which we deploy nova without keystone in the control
boxes.

Bug: T203177
Change-Id: I978f29c670ea5bdc770a243d57157fcb1810768a
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>