icinga::monitor::elasticsearch: throttle alerts notifications
Check that all shards are under a maximum size. This check is low frequency
since we can wait a number of days before taking action to reduce shard sizes.
We don't want to alert on shards going over limit for a short time,
since this can happen before segments are merged and space recovered.
The time before a shard failing and an alert being sent is "retry_interval * retries".

Bug: T206187
Change-Id: Idff73335dc1beac76b528ce5e866ccc174ddfc58