logstash: move to /srv/elasticsearch
Also fix the partman recipe to use standard /srv and fix the main raid1 to use all four disks.
Having all four disks partitioned the same and part of the raid makes it easier to replace disks
when needed.

This is also in preparation to hosting Kafka on logstash/elasticsearch nodes under /srv/kafka

Bug: T206454
Change-Id: I06eeacdefd808c0b23d18a207ee5d0a625933cba