Icinga: increase ripe atlas alerting threshold to 35
Currently about 370 hosts are part of the mesurement pool for IPv4 and
about 340 for IPv6 (for each destination).
We've been seeing a lot of alerts where we can't do much as the issue is
either along the path or on the source network.
Increasing the threshold to 35 (~10%) should make those alerts more
useful.

Bug: T205829
Change-Id: Ieb6d7200cf9bce82e0bfae418d27a9e13f46b372