base/icinga: use monitoring_hosts constant as NRPE allowed_hosts
The goal is to remove the hardcoded IP addresses from the default part
of the Hiera lookup and to add new server icinga1001.

Originally (I00e5b164ed4d760) i wanted to move the IP addresses
into Hiera itself, then i changed my mind to use the existing constant
MONTORING_HOSTS we already have for this.

This should avoid just repeating ourselves in a second place, also it
should protect against the issues we found like an IP of a former
Icinga server that was reused for something else but never removed
from this list.

Using DNS names should be fine in this place.

The list consists of the Icinga servers plus dbmonitor hosts who
would get the new ability as a side-effect. Though not strictly
needed it seems worth it instead of creating another list of hosts
with only Icinga servers.

Bug: T202782
Change-Id: I42f01a825c02165e86b059019e9e6ff9fc311479