role::eventlogging::analytics::files: reduce retention for archive
Due to an increase in events handled by Eventlogging, the disk
space consumption for /srv grew and now it is almost saturating
the related partition. Since these logs are rsynced to stat1005
and kept for 90 days, it is safe to apply a stricter retention
rule. The logs on stat1005 are not deleted by rsync, there is
a cron that uses `find -delete` running for files with mtime
older than 90d.

Bug: T206542
Change-Id: Ib0a378b7d7ddf1dfd484c0aa2b754d3c909d9ce9