statistics::rsync::eventlogging: reduce retention for archive
The current EL archive dir on stat1005 is ~1TB of compressed
logs. With the current increase in EL events handled this space
might grow a bit and end up using even more space, not a great
thing on stat1005 since the /srv dir has 15%/10% disk space$
available (oscillatting depending on the usage). The proposal is$
to prune logs after two months.$

Bug: T206542
Change-Id: Iffbd1e1eb9960a5a3442b601612fea4c5f597acc