Clean up removed rsyncd configs
The current puppetisation for rsyncd creates config snippets in
/etc/rsync.d which are then concatenated to a common rsyncd.conf.

But if an rsync::server::module is removed, the /etc/rsync.d/foo
file is never removed and thus continues to be used in rsyncd.conf.

Use recurse/purge for the snippet directory to fix this.

I'll also prepare some followup patches to convert rsyncd.conf to
a proper config files which then reads /etc/rsync.d via an include
of /etc/rsync.d.

Bug: T205618
Change-Id: I8f179469dee8751aca347e47681dbb40ca3cfbbc