Cleanup systemd state on Diamond removal
There's a bug when removing Diamond via the diamond::remove option:
The diamond.service remains in a failed state, as the service unit
cannot be found. Initially I hoped to fix this via
10a1d923228045afa7d6d6cb01c88e0f46b622d4, but but that turned out
to be not effective. Further debugging showed that this is actually
a bug in the diamond Debian package. I've filed it in Debian as
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=910787

We could this this in the Debian packages, but given that we're
removing it, I'm rather working around this in puppet, it's a
temporary measure anyway.

Bug: T183454
Change-Id: I005d6ca4dd515cc904af9d204e9cc70d325e9059