nagios_common: on jessie, also install libmonitoring-plugin-perl
libmonitoring-plugin-perl is replacing libnagios-plugin-perl
in stretch but is already available in jessie.

As suggested by Faidon on I3dcfb632cb32a4a9 install it here
as well to be able to make check_ssl work on both jessie and
stretch with the same code. Instead of having to add a debian
version detection in Perl in check_ssl itself.

Bug: T202782
Change-Id: I54017a2e720d768c3d441b974ce070390c956253