etcd::monitoring: replace Nagios::Plugin with Monitoring::Plugin on stretch
In stretch libnagios-plugin-perl will not exist anymore, so
this needs to be done before upgrading to stretch.

<s>In jessie both, libnagios-plugin-perl and libmonitoring-plugin-perl
exist, so this can be done anytime as long as all etcd is on jessie.</s>

(Adding an 'if stretch' nevertheless to make sure production isn't even
touched at all)

Doing this because i had to do the same thing with all other Perl
plugin based checks that are running on the Icinga server itself for
icinga 1001.

This is the only one that is using the Perl module and is executed via
NRPE on monitored hosts rather than the monitoring host, so just
doing this while at it.

Bug: T202782
Change-Id: I134af7b93ea85883b95c2509628688c1c610cbe5