interface::rps: support tg3 properly
Tigon3 is actually the most common ethernet chip in our production
fleet.  It's the primary interface on ~70% of the fleet and some
non-primary interface on another ~3%.

We're leaving the queues at their defaults (reasoning documented
in manifest) of 4 RX and 1 TX.  The interface-rps script is
updated to deal with the odd numbering of the queues (IRQ info has
off-by-one numbers in its labels, vs the sysfs view of the
queues).

Bug: T206105
Change-Id: Idb0aa495d26fdec2aa2ecebb2bd7a4d73def421a