interface::rps: always be NUMA aware
This removes the dependency on the weirdo global $numa_networking
we'd like to get rid of, so that interface::rps is always
NUMA-sensitive.  The only other consumer of $numa_networking is
the tlsproxy module's nginx config, which will have to be
re-examined separately.

The existing consumers of interface::rps are the cacheproxy hosts
(all of which have $numa_networking turned on) and the LVSes (all
of which do not), so we need to test this on an LVS or two first,
and be careful about deployment to the LVSes while pybal is
stopped (as it will blip their ethernet interfaces to change the
queue counts).

Change-Id: Idcfa124158f768769daf47bbe67ab610fd4488bd