toolforge: expand CSP allowed host list
There are a large number of 301 redirects and other magic in use that
currently work to get to various API endpoints that might be used by
tools. Forcing a cleanup of code linking to something non-canonical like
https://mediawiki.org/w/api.php is outside the scope of the CSP header,
so let's add all of the root domains that we are already wildcarding to
the list of allowed hosts.

Bug: T130748
Change-Id: Iabdf18a77a0d6c45ac35c61a94c51bf08e7689f5