profile::mediawiki::common: stop using class 'mediawiki'
The mediawiki class is full of anti-patterns as far as our puppet style
guide is concerned; it also makes it difficult to refactor the rest of
the MediaWiki classes.

So, remove it from profile::mediawiki::common. Also reduce the
requirements from a few maintenance classes that used to require
"mediawiki". Require correctly just "mediawiki::users" but in the future
we might end up passing the user as a paramter.

Change-Id: If8488bdca559e210dd553498051673f415bd879d