profile::analytics::refinery::job::camus: use camus to backup el-client-side
This new Camus job should allow Analytics to back up the
eventlogging-client-side topic on HDFS, freeing up some
space on various hosts (like eventlog1002) that keeps backup
of this data in case it needs to be refilled.

The job doesn't have any message decoder to grab the timestamp
from every event in the topic (and bucketize it accordingly) since
the message format is not json, but strings separated by tabs.
Camus should use its own wallclock timestamp to bucketize events
properly, but this is ok for the purpose of this task.

Bug: T206542
Change-Id: I09103d8de91fe63f1eb5fa7a3a7fffb30008586f