nagios_common: remove check_sslxNN
Nagios::Plugin is deprecated in stretch and needs to be replaced
with Monitoring::Plugin just like it has been done for
check_ssl in 2e919ffd2d23bedb1e.

But this command is not actually used anymore by anything, it has been
replaced by check_ssl_unified in Ie09fc1ac3e0de631.

So let's remove it to avoid somebody else trying to make this work
or wondering about it on stretch.

Bug: T202782
Change-Id: I1fc74d73cb454e4b52efbde229a7e606f2b3f23a