icinga: remove check_lonqueries.pl
Nagios::Plugin is deprecated in stretch and needs to be
replaced with Monitoring::Plugin just like it has been done
for check_ssl in 2e919ffd2d23bedb1e.

So this command will not work anymore on icinga1001.

But since i also don't see it actually being used, i am suggesting
to delete it rather than converting it.

Unless you still want to use this in the future.

Bug: T202782
Change-Id: Ie0053d72c795b7ca2dafb92c6d75ad2819d932d7