interface::rps: strict single CPU core per queue
In the original factoring of interface::rps, the aim of
dist_queues_to_cpus was to ensure (as close as possible) the even
distribution of queues to CPUs.

In cases like 20 queues vs 16 CPUs, the first 16 would be mapped
1:1, and the final 4 queues would be mapped to blocks of 4 CPUs
each.  In cases like 4 queues and 8 CPUs, each queue would be
mapped to 2 CPUs each.

With this patch, the extra 4 queues in the first case (20q/16c)
will simply map to the first 4 CPUs 1:1, giving 2 total queues to
each of those first 4 CPUs, and only one queue each to the
remaining 12 CPUs.  In the second example (4q/8c), the 4 queues
would map 1:1 to the first 4 CPUs, leaving the remaining 4 CPUs
queueless.

In the cases we've really cared about in the past, this was all
more or less esoteric anyways, as we usually use cards with very
large and configurable queue counts, and configured the card to
have exactly as many queues as CPUs.  We're now spreading this to
some cards with small fixed queues counts, so this all matters
more.

While having even distribution is nice, the downside of mapping a
single queue to fire on multiple cores is that packets from a
single flow (e.g. TCP connection) might end up on randomly
different cores within the set mapped for a given queue, yet share
per-flow data structures tracking loss, reassembly, etc which is
probably not very efficient at all in numerous ways.  Keeping a
single flow on a single core is probably worth considerably more
than the costs of uneven distribution.

Change-Id: I0e5711825c3d04e220568211c9b80a57199913d4