mcrouter: allow to tune timeouts until tko
The default settings are probably too sensitive
(3x1s timeouts before a shard is marked with TKO) for the
api/app servers clusters. Mcrouter does not behave like
nutcracker (remove the problematic shard, rehash and allow
to use another one while the other recovers) so failing super
fast after 3 timeouts might not be the right choice.
When mcrouter marks a shard as TKO, all the subsequent commands
return a failure (and we know that Mediawiki most of the times
tries to SET a value if the corresponding GET failed,
increasing the blast radius).

This change allows the configuration of the number of consecutive timeouts
to wait before marking the shard as TKO and raises it to 10
for api and appservers canaries.

Bug: T203786
Change-Id: I9cd5cb9def06b915f4667770fc4236d3ec738963