sonofgridengine: Correct and expand things enough to deploy a grid master
Since we've decided to stay on NFS to just get SGE deployed.  This touches
up and creates the needed elements to deploy a grid master using Son of
Grid Engine on Debian Stretch.  It won't work without more hiera stuff in
Horizon, but that's ok.  As long as it all points somewhere other than
Wikitech with conflicting namespaces.

Bug: T200557
Change-Id: Ibf6bbce6923e2a84630ed114a8d7022f334a02e0