eventlogging::server: rotate logs on size (not only on time)
The /var/log/eventlogging directory is usually stored on a tiny
root partition (few GBs), and it may happen that an increase in
error events (for example, due to a schema not validating) causes
a huge amounts of logs to be created, ending up in filling up the
disk partition over time. This patch should help alleviate this issue
adding a maximum size for the file log rotation.

Change-Id: I11bf06a1842ba1b6956803bedc8eb60ff75c3ad3