eventlogging: don't send logs to the syslog logfile
On eventlog1002 the /var/log partition (default for
daemon logging, even for eventbus) is on a relatively tiny
root partition (40G). A couple of days a go a deployment
caused a spike in events being logged with schema validation
errors, that caused the root partition to almost fill up since
syslog and daemon.log contains the same output as the
/var/log/eventlogging files, quadruplicating the size basically.

Since we don't need to log everything on syslog in this case,
it is better to log errors only in one place.

Change-Id: Ia514442eb5037f54da618c9e4871f09c09f862f2