mcrouter: add the probe_delay_initial_ms parameter
When a mcrouter gest 3 timeouts of 1 second from a memcached
shard, it waits probe_delay_initial_ms's default (10s) before
start sending health check probes (meant to verify if traffic
can be sent again or not). This means 10s of memcached errors
for mediawiki, that it definitely too long for our use case.

Bug: T203786
Change-Id: I600beab205e1e070d19ad598ca1a920fefee3b05