cloudvps: eqiad1: fix dmz_cidr adressing for inter-VM connections
A previous attempt to use this setting like this was reverted because we found
duplicated packets in the network. It turned out the duplicated packets were
not caused by this setting but by the neutron server duplicating IPs on restart
(different, unrelated stuff).

Let's reintroduce this setting and see what happens.

Bug: T206261
Change-Id: Ia0e5177262db85c67db28d1966988625b8df76c9
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>