icinga: avoid hardcoded nsca_user,nsca_group for NSCA service
On stretch the nsca_user can't be icinga, the nsca service
fails to start because it can't read the nsca config file.

This results in not having any passive check results from fundraising
hosts.

Avoid hardcoding the user name and group here, use our existing
variables we already use in all the other places now.

Because on jessie we have the mix of user/group = icinga/nagios
but on stretch we got rid of this, we need to keep the FIXME
in the erb template until einsteinium isn't production anymore.

Bug: T202782
Change-Id: I3fde93fb56433316c33bb38e6873b6749eb5eb5c