icinga/nsca: use systemd::service, unit file by systemd-sysv-generator
We want to replace service with systemd::service.
Only do it if on stretch, not touching production on jessie.
Add FIXME to remove that once off of stretch.

NSCA doesn't ship with a unit file, so for now we
are using the unit file that is generated by systemd.

The first time somebody runs /etc/init.d/nsca on the host,
the file /run/systemd/generator.late/nsca.service gets
generated. The contents of the template here are an exact
copy of that.

But at least this way we control the unit file here in puppet
and can build on it and it's more obvious what is even happening.

Had to use this command to even debug where the unit
file actually comes from that it tries to use, since nothing was
in puppet and it wasn't obvious systemd was used at all.

systemctl show -p FragmentPath nsca.service

In a future change we can write a better unit file and attempt
to get rid of the init.d script from package if we want to.

Bug: T202782
Change-Id: Ifbeca1be9d572cdd2e81fabe9a5433ccc2b564d6